Introduce tests to Calendars app
